+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 4.0.0 (2024-07-16)
+- add check for `switch-case` and `match` conditions
+- changed the error message for "Insane comparison ..."
+
 ### 3.2.0 (2023-06-09)
 - add check for `array ==(=) bool` conditions